+++ src/test/java/uk/gov/hmcts/cp/informantregister/e2e/ContractValidationDeadLetterIT.java
@@ -33,6 +33,7 @@ import uk.gov.hmcts.cp.informantregister.support.CapturedLog;
 import uk.gov.hmcts.cp.informantregister.support.PostgresTestSupport;
 import uk.gov.hmcts.cp.informantregister.support.ProcessedLogTestSupport;
 import uk.gov.hmcts.cp.informantregister.support.ServiceBusEmulatorTestSupport;
+import uk.gov.hmcts.cp.informantregister.support.ServiceTestSupport;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
@@ -91,6 +92,10 @@ class ContractValidationDeadLetterIT {
         registry.add("spring.datasource.username", PostgresTestSupport::username);
         registry.add("spring.datasource.password", PostgresTestSupport::password);
         registry.add("informantregister.servicebus.connection-string", () -> connectionString);
+        // CJSCPPUID. The service refuses to start without one, because a command sent
+        // anonymously is a command Results refuses; no suite here ever reaches Results.
+        registry.add("informantregister.results.system-user-id",
+                () -> ServiceTestSupport.SYSTEM_USER_ID);
     }
 
     @BeforeEach